Create dataHandling.py
This script serves two functions: first it compares data from the site-inspector scan (here 'inspect.csv') to the data from the registrar ('marchDom.csv') and then it reshapes the data into an output that can be more easily fed into a D3 visualization ('output.csv'). I would usually do this kind of script in R, so it's possible that the code can be cleaned up.